Added basic company valuation and scoring.

+++ src/dotnet/libs/Gimzo.Analysis/Fundamental/KeyMetrics.cs
@@ -15,14 +15,14 @@ public sealed class KeyMetrics
     public double? PriceEarningsToGrowthRate { get; init; }
     public decimal? BookValuePerShare { get; init; }
     public double? PriceToBookRatio { get; init; }
-    public double? Ebitda { get; init; }
+    public decimal? Ebitda { get; init; }
     public decimal? EnterpriseValue { get; init; }
     public double? DividendYield { get; init; }
     public double? DividendPayoutRatio { get; init; }
     public double? DebtToEquityRatio { get; init; }
     public decimal? CapitalExpenditures { get; init; }
     public decimal? FreeCashFlow { get; init; }
-    public decimal? ReturnOnEquity { get; init; }
+    public double? ReturnOnEquity { get; init; }
     public double? OneYearBeta { get; init; }
     public double? ThreeYearBeta { get; init; }
     public double? FiveYearBeta { get; init; }
